// src/database/prisma/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const CurrentShopType = {
  PRIMARY: 'PRIMARY',
  SECONDARY: 'SECONDARY'
} as const

export type CurrentShopType = (typeof CurrentShopType)[keyof typeof CurrentShopType]


export const PaymentType = {
  SUBSCRIPTION: 'SUBSCRIPTION',
  RENEWAL: 'RENEWAL',
  EXTRA_USER: 'EXTRA_USER',
  EXTRA_STORE: 'EXTRA_STORE'
} as const

export type PaymentType = (typeof PaymentType)[keyof typeof PaymentType]


export const Currency = {
  FCFA: 'FCFA',
  EUR: 'EUR',
  USD: 'USD'
} as const

export type Currency = (typeof Currency)[keyof typeof Currency]


export const PaymentStatus = {
  PENDING: 'PENDING',
  SUCCESS: 'SUCCESS',
  FAILED: 'FAILED',
  CANCELLED: 'CANCELLED'
} as const

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus]


export const PaymentMethod = {
  WAVE: 'WAVE',
  ORANGE_MONEY: 'ORANGE_MONEY',
  IN_APP: 'IN_APP'
} as const

export type PaymentMethod = (typeof PaymentMethod)[keyof typeof PaymentMethod]


export const SubscriptionStatus = {
  TRIAL: 'TRIAL',
  ACTIVE: 'ACTIVE',
  EXPIRED: 'EXPIRED',
  TRIAL_EXPIRED: 'TRIAL_EXPIRED',
  SUSPENDED: 'SUSPENDED'
} as const

export type SubscriptionStatus = (typeof SubscriptionStatus)[keyof typeof SubscriptionStatus]


export const FeatureCode = {
  ADVANCED_REPORTS: 'ADVANCED_REPORTS',
  TOP_PRODUCTS: 'TOP_PRODUCTS',
  LOW_STOCK_ALERT: 'LOW_STOCK_ALERT',
  STOCK_VALUE: 'STOCK_VALUE',
  SUPPLIER_MANAGEMENT: 'SUPPLIER_MANAGEMENT',
  EMPLOYEE_MANAGEMENT: 'EMPLOYEE_MANAGEMENT',
  ACCOUNTING: 'ACCOUNTING',
  MULTI_STORE: 'MULTI_STORE',
  EXPORT_EXCEL: 'EXPORT_EXCEL',
  EXPORT_PDF: 'EXPORT_PDF',
  API_ACCESS: 'API_ACCESS',
  OUT_OF_STOCK_ALERT: 'OUT_OF_STOCK_ALERT'
} as const

export type FeatureCode = (typeof FeatureCode)[keyof typeof FeatureCode]


export const PlanType = {
  FREE: 'FREE',
  BASIC: 'BASIC',
  PRO: 'PRO',
  PREMIUM: 'PREMIUM'
} as const

export type PlanType = (typeof PlanType)[keyof typeof PlanType]


export const UsageMetric = {
  SALES: 'SALES',
  PRODUCTS: 'PRODUCTS',
  USERS: 'USERS'
} as const

export type UsageMetric = (typeof UsageMetric)[keyof typeof UsageMetric]